requirements.txt
flask
gunicorn
psycopg2-binary
python-dotenv
# python-3.12.8